fix(release): restore macOS 14 archive compatibility
Keep the Command-1 handoff animation shortcut disabled outside dedicated debugging, disambiguate the animation opacity expression, and retain a macOS 14-compatible AVAssetExportSession path so the release archive can build.

+++ Hex/Features/Transcription/TranscriptionIndicatorView.swift
@@ -339,7 +339,7 @@ struct TranscriptionIndicatorView: View {
 
 			content
 				.frame(width: visibleCardSize.width, height: visibleCardSize.height)
-				.opacity(1 - min(1, handoffDepartureProgress * 4))
+				.opacity(Double(1 - min(CGFloat(1), handoffDepartureProgress * 4)))
 		}
 			.frame(width: panelSize.width, height: panelSize.height)
 	}
